requirements.txt
tiktoken>=0.4.0
python-telegram-bot>=20.3
openai>=0.27.6
azure-cognitiveservices-speech>=1.29.0
